+++ src/Paket.Core/PublicAPI.fs
@@ -371,9 +371,10 @@ type Dependencies(dependenciesFileName: string) =
         FindReferences.FindReferencesForPackage (PackageName package) |> this.Process
 
     // Packs all paket.template files.
-    member this.Pack(outputPath, ?buildConfig, ?version, ?releaseNotes, ?templateFile) =
+    member this.Pack(outputPath, ?buildConfig, ?version, ?releaseNotes, ?templateFile, ?workingDir) =
         let dependenciesFile = DependenciesFile.ReadFromFile dependenciesFileName
-        PackageProcess.Pack(dependenciesFile, outputPath, buildConfig, version, releaseNotes, templateFile)
+        let workingDir = defaultArg workingDir (dependenciesFile.FileName |> Path.GetDirectoryName)
+        PackageProcess.Pack(workingDir, dependenciesFile, outputPath, buildConfig, version, releaseNotes, templateFile)
 
     /// Pushes a nupkg file.
     static member Push(packageFileName, ?url, ?apiKey, (?endPoint: string), ?maxTrials) =